my-ai-soc-agent/requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-postgres>=0.0.17
python-nmap>=0.7.1
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
pgvector>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
paramiko>=3.4.0
streamlit>=1.38.0
watchdog>=4.0.0
redis>=5.0.0
